fix: support web-only extension for bundle resolution
