+++ development/rspack/utils.ts
@@ -30,6 +30,9 @@ export function createResolveExtensions({
       : []),
     // .ext.ts, .web.ts, .desktop.ts, .android.ts, .ios.ts, .native.ts
     ...['.ts', '.tsx', '.js', '.jsx'].map((ext) => `.${platform}${ext}`),
+    ...(platform === 'web' || platform === 'webEmbed'
+      ? ['.web-only.ts', '.web-only.tsx', '.web-only.mjs', '.web-only.js', '.web-only.jsx']
+      : []),
     '.web.ts',
     '.web.tsx',
     '.web.mjs',
